[ISSUE-160] Separa as consultaas em um arquivo a parte
